requirements.txt
aiogram==2.0
aiohttp==3.10.3
celery
fast_bitrix24==1.8.0
fastapi==0.112.0
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
uvicorn==0.30.6